requirements/production.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/production.in
#
amqp==5.1.1
    # via kombu
arrow==1.2.3
    # via -r requirements/production.in
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
attrs==22.2.0
    # via jsonschema
beautifulsoup4==4.12.2
    # via html-sanitizer
billiard==3.6.4.0
    # via celery
boto3==1.24.96
    # via -r requirements/production.in
botocore==1.27.96
    # via
    #   boto3
    #   s3transfer
celery[redis]==5.2.7
    # via
    #   -r requirements/production.in
    #   django-celery-beat
cffi==1.15.1
    # via cryptography
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cron-descriptor==1.2.35
    # via django-celery-beat
cryptography==40.0.1
    # via django-rest-knox
django==4.2
    # via
    #   -r requirements/production.in
    #   django-appconf
    #   django-celery-beat
    #   django-extensions
    #   django-filter
    #   django-health-check
    #   django-redis
    #   django-rest-knox
    #   django-storages
    #   djangorestframework
    #   drf-spectacular
    #   drf-standardized-errors
django-appconf==1.0.5
    # via django-imagekit
django-celery-beat==2.5.0
    # via -r requirements/production.in
django-environ==0.10.0
    # via -r requirements/production.in
django-extensions==3.2.1
    # via -r requirements/production.in
django-filter==23.1
    # via -r requirements/production.in
django-health-check==3.17.0
    # via -r requirements/production.in
django-imagekit==4.1.0
    # via -r requirements/production.in
django-redis==5.2.0
    # via -r requirements/production.in
django-rest-knox==4.2.0
    # via -r requirements/production.in
django-storages==1.13.2
    # via -r requirements/production.in
django-timezone-field==5.0
    # via django-celery-beat
djangorestframework==3.14.0
    # via
    #   -r requirements/production.in
    #   django-rest-knox
    #   drf-spectacular
    #   drf-standardized-errors
drf-spectacular==0.26.1
    # via
    #   -r requirements/production.in
    #   drf-standardized-errors
drf-standardized-errors[openapi]==0.12.5
    # via -r requirements/production.in
execnet==1.9.0
    # via pytest-xdist
factory-boy==3.2.1
    # via -r requirements/production.in
faker==18.4.0
    # via factory-boy
html-sanitizer==1.9.3
    # via -r requirements/production.in
inflection==0.5.1
    # via
    #   drf-spectacular
    #   drf-standardized-errors
iniconfig==2.0.0
    # via pytest
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.17.3
    # via drf-spectacular
kombu==5.2.4
    # via celery
lxml==4.9.2
    # via html-sanitizer
markdown-it-py==2.2.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mistune==2.0.5
    # via -r requirements/production.in
packaging==23.0
    # via
    #   pytest
    #   pytest-sugar
pilkit==2.0
    # via django-imagekit
pillow==9.5.0
    # via -r requirements/production.in
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.38
    # via click-repl
psycopg2-binary==2.9.6
    # via -r requirements/production.in
pycparser==2.21
    # via cffi
pygments==2.14.0
    # via rich
pyrsistent==0.19.3
    # via jsonschema
pytest==7.3.0
    # via
    #   -r requirements/production.in
    #   pytest-django
    #   pytest-sugar
    #   pytest-xdist
pytest-django==4.5.2
    # via -r requirements/production.in
pytest-sugar==0.9.6
    # via -r requirements/production.in
pytest-xdist==3.2.1
    # via -r requirements/production.in
python-crontab==2.7.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   arrow
    #   botocore
    #   faker
    #   python-crontab
pytz==2023.3
    # via
    #   celery
    #   django-timezone-field
    #   djangorestframework
pyuwsgi==2.0.21
    # via -r requirements/production.in
pyyaml==6.0
    # via drf-spectacular
redis==4.5.4
    # via
    #   celery
    #   django-redis
rich==13.3.3
    # via -r requirements/production.in
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   click-repl
    #   django-imagekit
    #   python-dateutil
soupsieve==2.4
    # via beautifulsoup4
sqlparse==0.4.3
    # via django
termcolor==2.2.0
    # via pytest-sugar
tzdata==2023.3
    # via django-celery-beat
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.15
    # via botocore
uwsgitop==0.11
    # via -r requirements/production.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
